refactor: slim entrypoint is the default
