+++ full/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "../cjs/full/index.js",
+  "module": "../esm/full/index.js",
+  "types": "../types/full/index.d.ts",
+  "sideEffects": false
+}
