--- slim/package.json
@@ -1,6 +1,0 @@
-{
-  "main": "../cjs/slim/index.js",
-  "module": "../esm/slim/index.js",
-  "types": "../types/slim/index.d.ts",
-  "sideEffects": false
-}
