+++ src/full/index.ts
@@ -2,7 +2,3 @@ export { SplitFactory } from './splitFactory';
 export { InLocalStorage } from '@splitsoftware/splitio-commons/src/storages/inLocalStorage/index';
 export { default as GoogleAnalyticsToSplit } from '@splitsoftware/splitio-commons/src/integrations/ga/GaToSplitPlugin';
 export { default as SplitToGoogleAnalytics } from '@splitsoftware/splitio-commons/src/integrations/ga/SplitToGaPlugin';
-export { ErrorLogger } from '@splitsoftware/splitio-commons/src/logger/browser/ErrorLogger';
-export { WarnLogger } from '@splitsoftware/splitio-commons/src/logger/browser/WarnLogger';
-export { InfoLogger } from '@splitsoftware/splitio-commons/src/logger/browser/InfoLogger';
-export { DebugLogger } from '@splitsoftware/splitio-commons/src/logger/browser/DebugLogger';
