+++ src/full/splitFactory.ts
@@ -1,12 +1,12 @@
-import { settingsValidator } from '../settings/slim';
-import { getModules } from '../platform/getModulesSlim';
+import { settingsValidator } from '../settings/full';
+import { getModules } from '../platform/getModulesFull';
 import { sdkFactory } from '@splitsoftware/splitio-commons/src/sdkFactory/index';
 import { ISdkFactoryParams } from '@splitsoftware/splitio-commons/src/sdkFactory/types';
 import { merge } from '@splitsoftware/splitio-commons/src/utils/lang';
 
 /**
  * SplitFactory with pluggable modules for Browser.
- * Doesn't include localhost mode and log messages.
+ * Includes localhost mode, verbose log messages and fetch ponyfill for old browsers.
  *
  * @param config configuration object used to instantiates the SDK
  * @param customModules optional object of SDK modules to overwrite default ones. Use with caution since, unlike `config`, this param is not validated.
