+++ src/platform/getFetchFull.ts
