+++ src/platform/getFetchSlim.ts
