+++ src/settings/full.ts
