--- src/slim/umd.ts
@@ -1,6 +1,0 @@
-// @ts-nocheck
-import { SplitFactory } from './splitFactory';
-
-SplitFactory.SplitFactory = SplitFactory;
-
-export default SplitFactory;
