+++ tsconfig.json
@@ -18,7 +18,7 @@
   },
   "include": [
     "src/index.ts",
-    "src/slim/index.ts"
+    "src/full/index.ts"
   ],
   "exclude": [
     "node_modules",
